[schema] Resolve search fields at schema parse time, share config between reference search and global search
