+++ packages/@sanity/default-layout/src/components/Search.js
@@ -51,10 +51,6 @@ function removeDupes(documents) {
     })
 }
 
-function is(typeName, type) {
-  return type.name === typeName || (type.type && is(typeName, type.type))
-}
-
 function search(query) {
   if (!client) {
     throw new Error('Sanity client is missing')
@@ -64,23 +60,9 @@ function search(query) {
     .filter(typeName => !typeName.startsWith('sanity.'))
     .map(typeName => schema.get(typeName))
 
-  const hasDocuments = candidateTypes.some(type => is('document', type))
-
-  const filterFn = hasDocuments
-    ? type => is('document', type)
-    : type => is('object', type) /* this is only for maintaining backwards compatibility after
-                                    introducing the 'document' type. Should be removed eventually
-                                  */
-
-  const searchableFields = flatten(
-    candidateTypes.filter(filterFn)
-      .map(type => (type.fields || [])
-        .filter(field => field.type.jsonType === 'string')
-        .map(field => field.name))
-  )
-
   const terms = query.split(/\s+/).filter(Boolean)
-  const uniqueFields = union(searchableFields)
+
+  const uniqueFields = union(flatten(candidateTypes.map(type => type.__unstable_searchFields)))
   const constraints = terms.map(term => uniqueFields.map(field => `${field} match '${term}*'`))
   const constraintString = constraints.map(constraint => `(${constraint.join(' || ')})`).join(' && ')
   return client.observable.fetch(`*[${constraintString}][0...10]`)
